refactor: introduced `CmdGenerator` and `DefaultCmdGenerator`
