+++ src/main/java/dev/jbang/source/CmdGenerator.java
@@ -1,0 +1,7 @@
+package dev.jbang.source;
+
+import java.io.IOException;
+
+public interface CmdGenerator {
+	String generate(Source src, RunContext ctx) throws IOException;
+}
